pyproject.toml
[tool.poetry]
name = "ful-split"
version = "0.1.0"
description = "tools for splitting radiohead's minidiscs"
authors = ["Benjamin Scharf <benjamin.d.scharf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
wget = "^3.2"
odfpy = "^1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
